K&R exercise 1.20 (Detab)
"Write a program detab that replaces tabs in the input with the proper
number of blanks to space to the next tab stop.  Assume a fixed set of
tab stops, say every n columns.  Should n be a variable or a symbolic
parameter."
Chose to pass tab width as an optional command-line argument, with a
default value provided at compile-time as a symbolic constant.